Most of a horrible "Linked List" implementation.
This is implemented in terms of SplDoublyLinkedList which just isn't going to cut it.
As nice as it is to have reduced memory footprint by using the SPL classes, it simply
does not expose any operations that make a linked list useful.

Currently, all inserts/removals are implemented by shifting values around as one would
for a vector.  At this stage I'm considering rewriting natively, though further performance
testing is needed.